--- .travis.yml
@@ -1,6 +1,0 @@
-language: go
-
-go:
-  - 1.0
-  - 1.1
-  - tip
